fix: message handler and script rule dependencies
